Reduce allocation cost reporting spans
This is done by leveraging Vert.x's Buffer
directly instead of relying on a
ByteArrayOutputStream
